refactor(ui): split browser panel into chrome renderers and a lifecycle controller
